wip - move some files around, trying to get closer to good boundaries
