+++ go/vt/vtctl/workflow/vexec/errors.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package workflow
+package vexec
 
 import "errors"
